requirements.txt
shellypy~=0.3.0
PyYAML